[libmce-glib] Add an example application
Serves as an example on how to use functionality provided by libmce-glib
and is useful also for testing and debugging the library itself.

Note that the application is built using libmce-glib headers and libraries
that are installed in the development environment - the same way regular
applications not included in libmce-glib source tree need to operate.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>